Revert "Each Qualifier can only appear once"
This reverts commit e25a0f5da3d2a85c338165c00a1a7434862ed8ff.

In that commit the Exists operator was ambiguous if there are two
qualifiers. Also, the approach taken in that commit would not
scale well as more qualifiers are added. There is another way to
implement this in the grammar (semantic predicates) but it would
require different syntax based on the programming language that
will use the grammar, and we'd like to keep the grammar language-
agnostic. Thus this patterning feature will need to be implemented
in the pattern-validator.